add rtt client support

+++ index.php
@@ -8,6 +8,9 @@ catch (Exception $e){
     exit;
 }
 $unm = "";
+$utoken = "";
+$uhtoken = "";
+
 if($loggingenabled && !touch($log)){
     $loggingenabled = FALSE;
 }
@@ -21,16 +24,20 @@ elseif($loggingenabled){
         logger("got incorrect log max lenth $tmp using default 5");
         unset($tmp);
     }
+    if(!is_int($maxlogsize)){
+        $maxlogsize = 1 * 1024 * 1024; //1MB
+    }
+    logger("=============================================");
 }
 
 function logger($data){
-    global $log,$loggingenabled;
+    global $log,$loggingenabled,$maxlogsize;
     if(!$loggingenabled){
         return;
     }
     $today = date("Y-m-d H:i:s");
     file_put_contents($log,"[$today]: $data\r\n",FILE_APPEND);
-    if(filesize($log) > 1* 1024 * 1024){ //1 MB
+    if(filesize($log) > $maxlogsize){ //set this in config. Default is 1 MB
         logrotate();
     }
 }
@@ -38,22 +45,35 @@ function logger($data){
 function logrotate(){
     global $logdir,$logfilename,$log,$loghistorylength;
     $logfiles = scandir($logdir);
-    $numlogs=1;
+    natsort($logfiles);
+    $logfiles = array_reverse($logfiles);
+    $postlog=array();
     foreach ($logfiles as $file) {
-        if(stripos($file, $logfilename) !== FALSE){
-            $numlogs++;
-            if($numlogs > $loghistorylength){
+        if(stripos($file, $logfilename) !== FALSE && $file != $logfilename){
+            $clognum = str_replace($logfilename.".","",$file);
+            $clognum = intval($clognum);
+            $postlog[] = "Current log is $clognum";
+            if($clognum >= $loghistorylength){
                 unlink("$logdir/$file");
+                $postlog[] = "removing log file $logdir/$file";
+            }
+            elseif($clognum == 0 || $clognum == "0"){
             }
             else{
-                rename("$logdir/$file", "$log.$numlogs");
+                rename("$logdir/$file", "$log.".($clognum+1));
+                $postlog[]= "renaming log from $logdir/$file to $log.".($clognum+1);
             }
         }
     }
     rename($log, "$log.1");
     touch($log);
+    foreach($postlog as $plogl){
+        logger($plogl);
+    }
 }
-
+/*
+//In my tests I found that the client identifier doesnt much matter.
+//Only the username and token needs to be changed
 function replaceCookie($value, $repl){
     $cookies = explode(";",$value);
     $toreplace = "";
@@ -71,9 +91,10 @@ function replaceCookie($value, $repl){
         return $value;
     }
 }
+*/
 
 function fixRequestUri($uri){
-    global $token, $username, $lbip, $destserver, $unm;//, $client
+    global $token, $username, $lbip, $destserver, $unm,$utoken;//, $client
     $uriparts = explode("?", $uri);
     if(count($uriparts)>1){
         $newuri = $uriparts[0]."?";
@@ -89,11 +110,14 @@ function fixRequestUri($uri){
         $kvp = explode("=", trim($query));
         switch ($kvp[0]) {
             case 'X-Plex-Token':
+                $utoken = $kvp[1];
+                logger("found user token $utoken");
                 $kvp[1] = $token;
                 logger("fixing url part ".$kvp[0]."=".$kvp[1]);
             break;
             case 'X-Plex-Username':
                 $unm = $kvp[1];
+                logger("found username $unm");
                 $kvp[1] = $username;
                 logger("fixing url part ".$kvp[0]."=".$kvp[1]);
             break;
@@ -124,30 +148,104 @@ function fixRequestUri($uri){
     return $newuri;
 }
 
+function getUsername($userheaders){
+    $ch = curl_init("https://plex.tv/users/account");
+    curl_setopt($ch,CURLOPT_ENCODING, '');
+    curl_setopt($ch,CURLOPT_RETURNTRANSFER,true);
+    curl_setopt($ch,CURLOPT_VERBOSE,1);
+    curl_setopt($ch, CURLOPT_HEADER, true);
+    curl_setopt($ch, CURLOPT_HTTPHEADER, $userheaders);
+    $output = curl_exec($ch);
+    curl_close($ch);
+    if(strpos($output,"<username>") != FALSE){
+        $st = strpos($output,"<username>") + strlen("<username>");
+        $ed = strpos($output,"</username>");
+        $usernm=substr($output,$st,($ed-$st));
+        logger("found username $usernm");
+        return $usernm;
+    }
+    else{
+        logger("Username tag not found");
+        return "";
+    }
+
+}
+
 $headers = array();
+$userheaders = array();
 $accepteh=FALSE;
 $acceptr=FALSE;
 
-logger($_SERVER['REQUEST_METHOD']);
-logger($_SERVER['REQUEST_URI']);
-$ruri = fixRequestUri($_SERVER['REQUEST_URI']);
-logger("http://$destserver:32400$ruri");
+
+
+//logger($_SERVER['REQUEST_METHOD']);
+//logger($_SERVER['REQUEST_URI']);
+$ouri = $_SERVER['REQUEST_URI'];
+$originalurl = "http://$destserver$ouri";
+if( stripos($ouri, "/:/prefs") !== FALSE ||
+    (stripos($ouri, "/:/plugins") !== FALSE && stripos($ouri, "/prefs") !== FALSE) ||
+    stripos($ouri, "/status/sessions") !== FALSE){
+    //block
+       // /:/plugins/<packagename>/prefs
+       // /:/prefs
+       // status/sessions
+    //allow
+      //  /video/<channel>
+      //  /music/<channel>
+      //  /photo/<channel>
+    logger("blocking $ouri");
+    logger("http://$destserver$ouri");
+    header("Location: $originalurl");
+    exit;
+}
+$ruri = fixRequestUri($ouri);
+$finalurl = "http://$destserver$ruri";
+//logger("http://$destserver$ruri");
 foreach (getallheaders() as $name => $value) {
+    $ovalue = $value;
     switch ($name) {
         case 'X-Plex-Token':
             logger("fixing $name");
+            $uhtoken = $value;
+            logger("found user token $utoken");
             $value = $token;
-        break;        
+            logger("adding $name $value");
+        break;
         case 'X-Plex-Username':
             logger("fixing $name");
             $unm = $value;
+            logger("found username $unm");
             $value = $username;
+            logger("adding $name $value");
         break;
         case 'Referer':
         case 'Host':
         case 'Origin':
             logger("fixing $name");
             $value = str_replace($lbip,$destserver, $value);
+            logger("adding $name $value");
+        break;
+        //With RTT Client you can instead stream media instead of redirecting
+        case 'Range':
+            if($enablerttclientmode && stripos($finalurl,"/library/parts/") !== FALSE){
+                logger("checking range $value");
+                $valsplit = explode("-",$value);
+                logger("checking range0 ".$valsplit[0]);
+                $valsplit[0] = explode("=",$valsplit[0])[1];
+                logger("checking range1 ".$valsplit[0]);
+                logger("checking range count ".count($valsplit));
+                if(count($valsplit) > 1 && intval($valsplit[1])-intval($valsplit[0])>$streamsize){
+                    logger('fixing range');
+                    $newend=intval($valsplit[0])+$streamsize;
+                    $value = "bytes=".$valsplit[0]."-$newend";
+                }
+                elseif(count($valsplit) == 1 || (count($valsplit) ==2 && $valsplit[1]=="") ){
+                    logger('fixing range');
+                    $newend=intval($valsplit[0])+$streamsize;
+                    $value = "bytes=".$valsplit[0]."-$newend";
+                }
+                logger("adding $name $value");
+            }
         break;
         /*
           //In my tests I found that the client identifier doesnt much matter.
@@ -163,13 +261,33 @@ foreach (getallheaders() as $name => $value) {
         //Detect gzip encoding
         $acceptr=TRUE;
     }
-    logger("adding $name $value");
+
     $headers[] = "$name: $value";
+    $userheaders[] = "$name: $ovalue";
 }
-logger("");
 
+//To enable RTT Client mode so that you can instead stream media instead of redirecting
+if(!$enablerttclientmode && (stripos($finalurl,"/library/parts/") !== FALSE || stripos($finalurl,"/video/:/transcode/") !== FALSE) ){
+    $finalurlre = "http://$destserverre$ruri";
+    header("Location: $finalurlre");
+    logger("Redirecting to redue load");
+    exit;
+}
 
-$ch = curl_init("http://$destserver:32400$ruri");
+if($unm == "" && ($utoken != "" || $uhtoken != "" ) ){
+    if($uhtoken == ""){
+       $uhtoken = $utoken;
+       $userheaders[] = "X-Plex-Token: $uhtoken";
+    }
+    elseif($utoken == ""){
+       $utoken = $uhtoken;
+    }
+    logger("looking up username with token $utoken");
+    $unm = getUsername($userheaders);
+    logger("changing username to $unm");
+}
+
+$ch = curl_init($finalurl);
 curl_setopt($ch,CURLOPT_ENCODING, '');
 curl_setopt($ch,CURLOPT_RETURNTRANSFER,true);
 curl_setopt($ch,CURLOPT_VERBOSE,1);
@@ -192,13 +310,13 @@ $firstrun = TRUE;
 
 
 
-logger($httpCode);
-logger($_SERVER['REQUEST_URI']);
-logger("http://$destserver:32400$ruri");
+//logger($httpCode);
+//logger($_SERVER['REQUEST_URI']);
+//logger("http://$destserver$ruri");
 foreach($header as $value){
     if($firstrun === TRUE){
         //The first header is always HTTP 1/1 CODE
-        //Sending this this messes things up
+        //Sending this messes things up
         $firstrun = FALSE;
         continue;
     }
@@ -210,17 +328,37 @@ foreach($header as $value){
         logger("Fixing ip");
         $value = str_replace($destserver, $lbip, $value);
     }
-    logger("returning $value");
     header("$value");
 }
-logger("trying to swap back username $unm");
-$body = str_replace('friendlyName="'.$destname.'"', 'friendlyName="'.$lbname.'"', $body);
-$body = str_replace('machineIdentifier="'.$destuuid.'"', 'machineIdentifier="'.$lbuuidkvp[$unm].'"', $body);
-$body = str_replace('myPlexUsername="'.$username.'"', 'myPlexUsername="'.$unm.'"', $body);
-logger("");
+
+
+if(strpos($body,'friendlyName="'.$destname.'"') !== FALSE){
+    $body = str_replace('friendlyName="'.$destname.'"', 'friendlyName="'.$lbname.'"', $body);
+    logger("trying to swap back friendlyname $lbname");
+}
+if(strpos($body,'machineIdentifier="'.$destuuid.'"') !== FALSE){
+    $body = str_replace('machineIdentifier="'.$destuuid.'"', 'machineIdentifier="'.$lbuuidkvp[$unm].'"', $body);
+    logger("trying to swap back machineID ".$lbuuidkvp[$unm]);
+}
+if(strpos($body,'myPlexUsername="'.$username.'"') !== FALSE){
+    $body = str_replace('myPlexUsername="'.$username.'"', 'myPlexUsername="'.$unm.'"', $body);
+    logger("trying to swap back username $unm");
+}
+if(strpos($body,'username="'.$username.'"') !== FALSE){
+    $body = str_replace('username="'.$username.'"', 'username="'.$unm.'"', $body);
+    logger("trying to swap back username $unm");
+}
+if(strpos($body,'authToken="'.$token.'"') !== FALSE){
+    $body = str_replace('authToken="'.$token.'"', 'authToken="'.$utoken.'"', $body);
+    logger("trying to swap back token $utoken");
+}
+
 if(($accepteh === TRUE && $acceptr === TRUE)){
     echo gzencode($body);
 }
 else{
     echo $body;
 }
+
+exit;
+?>
